+++ apps/web/src/utils/wagmi.ts
@@ -11,7 +11,7 @@ import { InjectedConnector } from 'wagmi/connectors/injected'
 import { LedgerConnector } from 'wagmi/connectors/ledger'
 import { MetaMaskConnector } from 'wagmi/connectors/metaMask'
 import { jsonRpcProvider } from 'wagmi/providers/jsonRpc'
-import { DEFAULT_APP_METADATA } from 'views/Notifications/utils/constants'
+import { DEFAULT_APP_METADATA } from 'views/Notifications/constants'
 import { WalletConnectConnector } from './WalletConnectConnector2'
 
 // get most configs chain nodes length
