+++ apps/web/src/views/Notifications/components/PushSubscribeButton/PushSubscribeButton.tsx
@@ -59,7 +59,6 @@ export default function PushSubscriptionButton({
       <CommitButton
         variant="primary"
         onClick={handleSubscriptionAction}
-        //   disabled={Boolean(errorText)}
         isLoading={isSubscribing || isUnsubscribing}
         height="55px"
       >
