+++ apps/web/src/views/Notifications/types.ts
@@ -12,3 +12,13 @@ export type NotificationType = {
   title: string
   type: string
 }
+
+export type NotifyType = {
+  title: string
+  description: string
+}
+
+export type RelayerType = {
+  value: string | undefined
+  label: string
+}
