+++ apps/web/src/views/Pools/components/LockedPool/hooks/useLockedPool.tsx
@@ -56,7 +56,7 @@ export default function useLockedPool(hookArgs: HookArgs): HookReturn {
   const handleDeposit = useCallback(
     async (convertedStakeAmount: BigNumber, lockDuration: number) => {
       const callOptions = {
-        gas: vaultPoolConfig[VaultKey.CakeVault].gasLimit,
+        gasLimit: vaultPoolConfig[VaultKey.CakeVault].gasLimit,
       }
 
       const receipt = await fetchWithCatchTxError(() => {
