+++ apps/web/src/views/Pools/hooks/useUnstakePool.ts
@@ -5,7 +5,7 @@ import { useSousChef } from 'hooks/useContract'
 import { useGasPrice } from 'state/user/hooks'
 
 const options = {
-  gas: DEFAULT_GAS_LIMIT,
+  gasLimit: DEFAULT_GAS_LIMIT,
 }
 
 const sousUnstake = (sousChefContract: any, amount: string, decimals: number, gasPrice: bigint) => {
